chore: remove models types and organize better
