+++ packages/types/index.ts
@@ -1,7 +1,1 @@
-export * from './model/user';
-export * from './model/student';
-export * from './model/reaction';
-export * from './model/comment';
-export * from './model/enrollment';
-export * from './model/disciplina';
-export * from './user/oauth2';
+export * from './src/oauth2';
